Revert "Update gitlab4j to 5.x (#976)"
This reverts commit 15faa6153560173c0ab7f1fdd37136e1b6caa5f2.

+++ legend-sdlc-server/src/test/java/org/finos/legend/sdlc/server/gitlab/api/GitLabEntityApiTestResource.java
@@ -173,8 +173,8 @@ public class GitLabEntityApiTestResource
 
         GitLabProjectId sdlcGitLabProjectId = GitLabProjectId.parseProjectId(projectId);
         MergeRequestApi mergeRequestApi = this.gitLabMemberUserContext.getGitLabAPI().getMergeRequestApi();
-        long parsedMergeRequestId = Long.parseLong(reviewId);
-        long gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
+        int parsedMergeRequestId = Integer.parseInt(reviewId);
+        int gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
 
         String requiredStatus = "can_be_merged";
         CallUntil<MergeRequest, GitLabApiException> callUntil = CallUntil.callUntil(
@@ -320,8 +320,8 @@ public class GitLabEntityApiTestResource
 
         GitLabProjectId sdlcGitLabProjectId = GitLabProjectId.parseProjectId(projectId);
         MergeRequestApi mergeRequestApi = this.gitLabMemberUserContext.getGitLabAPI().getMergeRequestApi();
-        long parsedMergeRequestId = Long.parseLong(reviewId);
-        long gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
+        int parsedMergeRequestId = Integer.parseInt(reviewId);
+        int gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
 
         String requiredStatus = "can_be_merged";
         CallUntil<MergeRequest, GitLabApiException> callUntil = CallUntil.callUntil(
@@ -473,8 +473,8 @@ public class GitLabEntityApiTestResource
 
         GitLabProjectId sdlcGitLabProjectId = GitLabProjectId.parseProjectId(projectId);
         MergeRequestApi mergeRequestApi = this.gitLabMemberUserContext.getGitLabAPI().getMergeRequestApi();
-        long parsedMergeRequestId = Long.parseLong(reviewId);
-        long gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
+        int parsedMergeRequestId = Integer.parseInt(reviewId);
+        int gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
 
         String requiredStatus = "can_be_merged";
         CallUntil<MergeRequest, GitLabApiException> callUntil = CallUntil.callUntil(
@@ -626,8 +626,8 @@ public class GitLabEntityApiTestResource
 
         GitLabProjectId sdlcGitLabProjectId = GitLabProjectId.parseProjectId(projectId);
         MergeRequestApi mergeRequestApi = this.gitLabMemberUserContext.getGitLabAPI().getMergeRequestApi();
-        long parsedMergeRequestId = Long.parseLong(reviewId);
-        long gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
+        int parsedMergeRequestId = Integer.parseInt(reviewId);
+        int gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
 
         String requiredStatus = "can_be_merged";
         CallUntil<MergeRequest, GitLabApiException> callUntil = CallUntil.callUntil(
